ScoreArrrangementモジュールを削除し、Frameインスタンスのまとまりを作成するためのFrameFactoryクラスを作成
